Clear vectors when using CRC32 hash
If a vector is reused with a CRC32 hash, then the new hash gets appended to the old hash. This is against the intuition of a user, who expects a new hash where the old value is overwritten. This commit fixes the issue by clearing the hash vector before appending the new hash bytes.

Reported-by: Lars Nielsen <lani@eng.au.dk>
Fixes: 4e28e8ad48d18a3ca0c09bb5eaca39654398d720
Tested-by: ChrisG55 <christian_g_55@msn.com>
Signed-off-by: ChrisG55 <christian_g_55@msn.com>